requirements.txt
Django==5.2.3
Pillow==10.2.0
djangorestframework==3.16.0 